requirements.txt
discord.py
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
python-dotenv